Added docker-compose.yml, datasource for postgres and controllers,repos
With this commit I added the docker-compose.yml file so that we can create the postgres container along with the postgres admin dashboard to be able to see and edit the database's data. I also deleted the application.properties file and replaced it with the .yml one and added in basic configuration for datasource & jpa. I also created the basic controller, model & the repo for the user api.