First users endpoint created: createdAll
This createAll endpoint creates massively users. It gets the images for the users from the url "https://picsum.photos/300/300?random". The 300/300 are the images size. What must be explained is the httpclient. It uses the method setRedirectStrategy(new LaxRedirectStrategy()).build() since we need to redirect to get the actual image(The url "https://picsum.photos/300/300?random" redirects us to a new url which has the image). Once that is completed we save the image.

Next steps: With the library Faker get data for the firstname, lastname, location & along with the images save the user to our db.